Add library build
Every library module should output a library that
can be used by other libraries or apps.
The dependencies must not be included but must be treated as
externals. The provided function adds every dependency as external
automatically and sets the necessary library flags.
During publishing, the entry point in the package.json is changed to the
built library. That way, the library works as usual during development
and is not required to be built explicitly.

Build d.ts files even though package contains source code
to make code completion work in IntelliJ
https://youtrack.jetbrains.com/issue/WEB-29638/Pure-ts-node-modules-do-not-work-as-expected-with-auto-import-complete